moving forward with burn implementation
