+++ r2l-agents/src/lib.rs
@@ -3,21 +3,19 @@ pub mod ppo;
 pub mod vpg;
 
 use r2l_core::{
-    agents::Agent2, distributions::DistributionKind,
-    policies::learning_modules::LearningModuleKind, utils::rollout_buffer::RolloutBuffer,
+    agents::Agent, distributions::DistributionKind, policies::learning_modules::LearningModuleKind,
+    utils::rollout_buffer::RolloutBuffer,
 };
 
-use crate::{a2c::a2c3::A2C3, ppo::ppo3::PPO3, vpg::vpg3::VPG3};
-// use candle_core::Result;
-// use r2l_core::{agents::Agent, policies::PolicyKind, utils::rollout_buffer::RolloutBuffer};
+use crate::{a2c::A2C, ppo::PPO, vpg::VPG};
 
 pub enum AgentKind {
-    A2C(A2C3<DistributionKind, LearningModuleKind>),
-    PPO(PPO3<DistributionKind, LearningModuleKind>),
-    VPG(VPG3<DistributionKind, LearningModuleKind>),
+    A2C(A2C<DistributionKind, LearningModuleKind>),
+    PPO(PPO<DistributionKind, LearningModuleKind>),
+    VPG(VPG<DistributionKind, LearningModuleKind>),
 }
 
-impl Agent2 for AgentKind {
+impl Agent for AgentKind {
     type Dist = DistributionKind;
 
     fn learn(&mut self, rollouts: Vec<RolloutBuffer>) -> candle_core::Result<()> {
