+++ r2l-api/tests/ppo.rs
@@ -1,7 +1,7 @@
-use candle_core::{Device, Result};
+use candle_core::Result;
 use r2l_api::builders::on_policy_algo::OnPolicyAlgorithmBuilder;
 use r2l_api::test_utils::run_gym_episodes;
-use r2l_core::agents::Agent2;
+use r2l_core::agents::Agent;
 use r2l_core::{Algorithm, on_policy_algorithm::LearningSchedule};
 
 const NUM_ENVIRONMENTS: usize = 10;
