+++ r2l-core/src/on_policy_algorithm.rs
@@ -1,5 +1,5 @@
 use crate::Algorithm;
-use crate::agents::Agent2;
+use crate::agents::Agent;
 use crate::env::{EnvPool, RolloutMode};
 use crate::utils::rollout_buffer::RolloutBuffer;
 use candle_core::Result;
@@ -87,7 +87,7 @@ impl OnPolicyHooks {
     }
 }
 
-pub struct OnPolicyAlgorithm2<E: EnvPool, A: Agent2> {
+pub struct OnPolicyAlgorithm2<E: EnvPool, A: Agent> {
     pub env_pool: E,
     pub agent: A,
     pub learning_schedule: LearningSchedule,
@@ -95,7 +95,7 @@ pub struct OnPolicyAlgorithm2<E: EnvPool, A: Agent2> {
     pub hooks: OnPolicyHooks,
 }
 
-impl<E: EnvPool, A: Agent2> OnPolicyAlgorithm2<E, A> {
+impl<E: EnvPool, A: Agent> OnPolicyAlgorithm2<E, A> {
     pub fn new(
         env_pool: E,
         agent: A,
@@ -149,7 +149,7 @@ impl<E: EnvPool, A: Agent2> OnPolicyAlgorithm2<E, A> {
     }
 }
 
-impl<E: EnvPool, A: Agent2> Algorithm for OnPolicyAlgorithm2<E, A> {
+impl<E: EnvPool, A: Agent> Algorithm for OnPolicyAlgorithm2<E, A> {
     fn train(&mut self) -> Result<()> {
         if self.hooks.call_before_training_hook()? {
             return Ok(());
