+++ r2l-core2/src/distributions/mod.rs
@@ -1,12 +1,16 @@
 use crate::env::{Action, Logp, Observation};
 use std::{f32, fmt::Debug};
 
-pub trait Distribution<O: Observation, A: Action, L: Logp>: Sync + Debug + 'static {
+pub trait Distribution: Sync + Debug + 'static {
+    type Observation: Observation;
+    type Action: Action;
+    type Logp: Logp;
+
     // get action for a single observation
-    fn get_action(&self, observation: O) -> (A, f32);
+    fn get_action(&self, observation: Self::Observation) -> (Self::Action, f32);
 
     // get logps for more than one states/action pairs
-    fn log_probs(&self, states: &[O], actions: &[A]) -> Vec<L>;
+    fn log_probs(&self, states: &[Self::Observation], actions: &[Self::Action]) -> Vec<Self::Logp>;
 
     // TODO: re evaluate this
     fn std(&self) -> f32;
