+++ r2l-core2/src/lib.rs
@@ -30,6 +30,6 @@ pub mod env;
 pub mod env_pools;
 pub mod numeric;
 pub mod on_policy_algorithm;
-pub mod policies;
+// pub mod policies;
 pub mod thread_safe_sequential;
 pub mod utils;
