+++ r2l-policies/src/paralell_actor_critic.rs
@@ -1,9 +1,8 @@
 use burn::tensor::{Tensor, backend::AutodiffBackend};
-use r2l_core2::{
-    policies::{Policy, PolicyWithValueFunction},
-    thread_safe_sequential::ThreadSafeSequential,
-};
+use r2l_core2::thread_safe_sequential::ThreadSafeSequential;
 use r2l_distributions::DistribnutionKind;
+
+use crate::LearningModule;
 
 pub struct ParalellActorCritic<B: AutodiffBackend> {
     distribution: DistribnutionKind<B>,
@@ -16,24 +15,15 @@ pub struct ParalellActorCriticLosses<B: AutodiffBackend> {
     value_loss: Tensor<B, 2>,
 }
 
-impl<B: AutodiffBackend> Policy for ParalellActorCritic<B> {
-    type Obs = Tensor<B, 2>;
-    type Act = Tensor<B, 2>;
-    type Logp = Tensor<B, 2>;
-    type Losses = ();
-    type Dist = DistribnutionKind<B>;
-
-    fn distribution(&self) -> Self::Dist {
-        self.distribution.clone()
-    }
-
-    fn update(&mut self, losses: Self::Losses) {
-        todo!()
-    }
+pub struct PolicyValueLosses<B: AutodiffBackend> {
+    pub policy_loss: Tensor<B, 1>,
+    pub value_loss: Tensor<B, 1>,
 }
 
-impl<B: AutodiffBackend> PolicyWithValueFunction for ParalellActorCritic<B> {
-    fn calculate_value(&self, observation: <Self as Policy>::Obs) -> f32 {
+impl<B: AutodiffBackend> LearningModule for ParalellActorCritic<B> {
+    type Losses = PolicyValueLosses<B>;
+
+    fn update(&mut self, losses: Self::Losses) {
         todo!()
     }
 }
